Implemented Arrival/Departures selection
Many changes, the StationView class is now mostly useful.
It is possible to select arrivals/departures from the menu to change the
displayed section.

Implemented using the QWebElement dom functions to annotate the elements
as departures or arrivals, as the structure of the viaggiatreno site is
unsuitable for a css only solution.